# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Skolr school super app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: school_id
          in: query
          schema:
            type: string
        - name: class_id
          in: query
          schema:
            type: string
        - name: parent_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  # ── CLASSES ──────────────────────────────────────────────────────────────
...
  /timetable:
    get:
      operationId: listTimetableEntries
      tags: [timetable]
      parameters:
        - name: student_id
          in: query
          schema:
            type: string
        - name: class_id
          in: query
          schema:
            type: string
        - name: teacher_id
          in: query
          schema:
            type: string
        - name: day_of_week
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"
    post:
      operationId: createTimetableEntry
      tags: [timetable]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimetableEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
  /timetable/{id}:
    patch:
      operationId: updateTimetableEntry
      tags: [timetable]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimetableEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
    delete:
      operationId: deleteTimetableEntry
      tags: [timetable]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── EVENTS ────────────────────────────────────────────────────────────────
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      parameters:
        - name: student_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ── SCHOOLS ───────────────────────────────────────────────────────────────
...
  /parent-student-links:
    get:
      operationId: listParentStudentLinks
      tags: [parent-student-links]
      summary: List parent-student links
      parameters:
        - name: parent_user_id
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParentStudentLink"
    post:
      operationId: createParentStudentLink
      tags: [parent-student-links]
      summary: Link a parent to a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParentStudentLinkInput"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentStudentLink"
  /parent-student-links/{id}:
    delete:
      operationId: deleteParentStudentLink
      tags: [parent-student-links]
      summary: Remove a parent-student link
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, profile_id, class_id, school_id, grade, created_at]
      properties:
        id:                  { type: string }
        profile_id:          { type: string }
        class_id:            { type: ["string", "null"] }
        school_id:           { type: string }
        grade:               { type: string }
        date_of_birth:       { type: ["string", "null"] }
        student_number:      { type: ["string", "null"] }
        full_name:           { type: string }
        avatar_url:          { type: ["string", "null"] }
        created_at:          { type: string }
    StudentInput:
      type: object
      required: [profile_id, school_id, grade]
      properties:
        profile_id:      { type: string }
        class_id:        { type: string }
        school_id:       { type: string }
        grade:           { type: string }
        date_of_birth:   { type: string }
        student_number:  { type: string }
    StudentUpdate:
      type: object
      properties:
        class_id:      { type: string }
        grade:         { type: string }
        date_of_birth: { type: string }
...
    TimetableEntry:
      type: object
      required: [id, subject_id, teacher_id, day_of_week, start_time, end_time, type, school_id, created_at]
      properties:
        id:           { type: string }
        student_id:   { type: ["string", "null"] }
        class_id:     { type: ["string", "null"] }
        subject_id:   { type: string }
        teacher_id:   { type: string }
        day_of_week:  { type: string }
        start_time:   { type: string }
        end_time:     { type: string }
        room:         { type: ["string", "null"] }
        type:         { type: string }
        school_id:    { type: string }
        subject_name: { type: ["string", "null"] }
        teacher_name: { type: ["string", "null"] }
        class_name:   { type: ["string", "null"] }
        created_at:   { type: string }
    TimetableEntryInput:
      type: object
      required: [subject_id, teacher_id, day_of_week, start_time, end_time, type, school_id]
      properties:
        student_id:  { type: string }
        class_id:    { type: string }
        subject_id:  { type: string }
        teacher_id:  { type: string }
        day_of_week: { type: string }
        start_time:  { type: string }
        end_time:    { type: string }
        room:        { type: string }
        type:        { type: string }
        school_id:   { type: string }
    TimetableEntryUpdate:
      type: object
      properties:
        day_of_week: { type: string }
        start_time:  { type: string }
        end_time:    { type: string }
        room:        { type: string }
        type:        { type: string }
...
    Report:
      type: object
      required: [id, student_id, title, term, year, file_url, school_id, created_at]
      properties:
        id:           { type: string }
        student_id:   { type: string }
        title:        { type: string }
        term:         { type: string }
        year:         { type: integer }
        file_url:     { type: string }
        visible_to_student: { type: boolean }
        school_id:    { type: string }
        student_name: { type: ["string", "null"] }
        created_at:   { type: string }
    ReportInput:
      type: object
      required: [student_id, title, term, year, file_url, school_id]
      properties:
        student_id:         { type: string }
        title:              { type: string }
        term:               { type: string }
        year:               { type: integer }
        file_url:           { type: string }
        visible_to_student: { type: boolean }
        school_id:          { type: string }
...
    ParentStudentLink:
      type: object
      required: [id, parent_user_id, student_id, school_id, created_at]
      properties:
        id:             { type: string }
        parent_user_id: { type: string }
        student_id:     { type: string }
        school_id:      { type: string }
        student_name:   { type: ["string", "null"] }
        student_grade:  { type: ["string", "null"] }
        created_at:     { type: string }
    ParentStudentLinkInput:
      type: object
      required: [parent_user_id, student_id, school_id]
      properties:
        parent_user_id: { type: string }
        student_id:     { type: string }
        school_id:      { type: string }
...