# .migration-backup/lib/api-zod/src/generated/types/timetableEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skolr school super app API
 * OpenAPI spec version: 0.1.0
 */

export interface TimetableEntry {
  id: string;
  /** @nullable */
  student_id?: string | null;
  /** @nullable */
  class_id?: string | null;
  subject_id: string;
  teacher_id: string;
  day_of_week: string;
  start_time: string;
  end_time: string;
  /** @nullable */
  room?: string | null;
  type: string;
  school_id: string;
  /** @nullable */
  subject_name?: string | null;
  /** @nullable */
  teacher_name?: string | null;
  /** @nullable */
  class_name?: string | null;
  created_at: string;
}