# .migration-backup/lib/api-zod/src/generated/types/timetableEntryUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skolr school super app API
 * OpenAPI spec version: 0.1.0
 */

export interface TimetableEntryUpdate {
  day_of_week?: string;
  start_time?: string;
  end_time?: string;
  room?: string;
  type?: string;
}