# lib/api-zod/src/generated/types/parentStudentLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Skolr school super app API
 * OpenAPI spec version: 0.1.0
 */

export interface ParentStudentLink {
  id: string;
  parent_user_id: string;
  student_id: string;
  school_id: string;
  /** @nullable */
  student_name?: string | null;
  /** @nullable */
  student_grade?: string | null;
  created_at: string;
}